Masonry
Masonry is successfully working and displaying a grid layout with
different sized images.

The container, however, still needs to have a scalable width applied to
it, so it will resize with the width of the window. Masonry may support
this with the “bindresize” method. Otherwise, this SmartResize jsfiddle
may be adequate to solve the problem:
http://jsfiddle.net/matthewabrman/6R2DU/